feat: Alter SearchVectorFast function signature to return int32

+++ go_api/engine/engine.go
@@ -1,4 +1,4 @@
 package engine
 
 //go:noescape
-func SearchVectorFast(q *float32, scratch *byte) float32
+func SearchVectorFast(q *float32, scratch *byte) int32
